Zmiana nazwy klas Map -> Repo
